setup for the Spiegelman 2016 paper
The changes to core.cc are probably not necessary any more; I used them
for the boundary conditions, but the same thing could now probably be
done by prescribing the veolcity boundary consitions as function x:...
(and leaving the y velocity open).

The changes to the assembly.cc are necessary for testing the influence
of the compressibility, otherwise the dynamic effect of compressibility
is not included in the equations. This is kind of a hack right now, but
with the old version the velocity divergence would not change due to
changes in the dynamic pressure, and that is exactly the effect that I
suspect is important here.